refactor: extract camera-ready report writer for issue 3385
